+++ previews/elements/progress_bar_component_preview.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Elements::ProgressBarComponentPreview < ViewComponent::Preview
+  def default
+    render Bulma::ProgressBarComponent.new 90
+  end
+
+  def colors
+  end
+
+  def sizes
+  end
+end