+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
-- Created new component `ProgressBarComponent`
+- Created previews for `Bulma::ProgressBarComponent`
+- Created new component `Bulma::ProgressBarComponent`
 - Created previews for `Bulma::ButtonsComponent`
 - Created new component `Bulma::ButtonsComponent`
 - Updated `Elements::ButtonComponentPreview` to pass content as parameter